CI: disable prometheus
Disable Prometheus in CI until we're actually using it in some meaninful
way. At this time, it is lacking in proper configuration limiting memory,
scraping, and more.

On average, these Pods are using in excess of 2Gi of ram before being evicted,
causing failures on various nodes.

Until such time as the following are addressed, disable it.

Relates to gitlab-org/charts/gitlab#330

Relates to gitlab-org/charts/gitlab#393