Allow configuring multiple Redis instances
From #1644

Allow configuration of multiple Redis instances, for splitting
persistence classes.

This implements the necessary configuration items, and templating of
configuration and secrets.

Currently, it is expected that if this configuration is used, all
Redis instances will be external to this Helm chart. This means that
we expect `redis.install` to be false, and all necessary queues to be
individuall configured under `global.redis`.

Side effects of this change:

- ignorance of `.Values.redis`, in favor of `.Values.global.redis`
- DRY-ification of `resque.yml` across all GitLab charts to a central
template (`gitlab.rails.redis.resque`)
- Secret mounting blocks are now rendered via `gitlab.redis.secrets`,
which in turn consumes `gitlab.redis.secret`
- Base implementation of ActionCable's `cable.yml`

Questions remaining:

- Does `gitlab/geo-logcursor` need all of these configurations?
- Does `gitlab/gitlab-exporter` actually access Redis as configured?

Documentation to be done:

- Add additional Redis properties to documentation
- Move Redis configuration documentation our of Unicorn
- Add documentation of multiple Redis to doc/advanced/external-redis