CI: disable pipeline jobs for auto-deploy branches on .com
Disable CI jobs for pipelines caused by auto-deploy branches on GitLab.com

The various branches created and managed by the Delivery team and the
auto-deploy processes do not need to be deployed into out CI clusters,
where they consume resources. They'll not be used by the teams (as far
as I could confirm), are based off our master branches, which should be
kept green at all times as it is.

Closes #2006