Redis: Fix templating of secret when password.enabled=false
Address #2051 by moving the `if` condition within `gitlab.redis.secret`
This is accomplished by including `gitlab.redis.configMerge`, which
is safe to re-use because of the gating within that function, having
no real impact on execution time.

This also ensure the complication of the if-case stays outside of the
caller.