Make gitlab.boolean.local return an empty string for false
Because this is a template, it can only return a string. The string
'false' is truthy because it's not empty. This template now returns:

- 'true' for true.
- '' for false.

This means that the returned values are usable in truthiness checks with
`default`, and also by the Docker image itself (which only compares
directly with 'true' anyway).