Database: support load balancing reads in Rails
Add support for the configuration of read-only query load balancing
as a part of Rails.

Implements (at a global level), the support for configuration of
https://docs.gitlab.com/ee/administration/database_load_balancing.html

Added:
- checkConfig to ensure that internal PostgreSQL is not being installed
- `global.psql.load_balancing` properties.
- RSpec to test checkConfig
- RSpec to test configuration renders as expected.
- Examples under `examples/database/values-loadbalancing-*`

Tested configuration

NOTE: Properties were added here in `snake_case` to directly align to
upstream documentation of these properties. They _can_ be swapped in
the future, but it may be time to re-evaluate that choice.

Closes #1004